requirements.txt
flask>=3.0.0
PyPDF2>=3.0.0
python-pptx>=0.6.21
